rename shifts module to scheduling
