+++ pyworkforce/scheduling/__init__.py
@@ -1,0 +1,3 @@
+from pyworkforce.scheduling.shifts_selection import MinAbsDifference, MinRequiredResources
+
+__all__ = ["MinAbsDifference", "MinRequiredResources"]
