+++ pyworkforce/scheduling/utils.py
