--- pyworkforce/shifts/__init__.py
@@ -1,3 +1,0 @@
-from pyworkforce.shifts.shifts_selection import MinAbsDifference, MinRequiredResources
-
-__all__ = ["MinAbsDifference", "MinRequiredResources"]
